+++ android/src/main/kotlin/dev/steenbakker/mobile_scanner/MobileScannerHandler.kt
@@ -42,6 +42,13 @@ class MobileScannerHandler(
 
     private val analyzeImageSuccessCallback: AnalyzerSuccessCallback = {
         Handler(Looper.getMainLooper()).post {
+            // TODO: Open for discussion if we want to publish the results on the barcode stream as well.
+//            // Also publish on controller result
+//            barcodeHandler.publishEvent(mapOf(
+//                "name" to "barcode",
+//                "data" to it,
+//            ))
+
             analyzerResult?.success(mapOf(
                 "name" to "barcode",
                 "data" to it